requirements.txt
einops==0.2.0
gdown==4.4.0
prefetch_generator==1.0.1
imgaug==0.4.0
lmdb>=1.0.0
scikit-image>=0.21.0
pyyaml
torch
torchvision
opencv-python-headless
numpy==1.23.5
Pillow==9.4.0
pandas
